Feature: Validating before deleting organization permission if the user is the last admin of the organization

+++ connect/usecases/authorizations/delete.py
@@ -2,17 +2,20 @@ from django.db.models import QuerySet
 from rest_framework.exceptions import PermissionDenied
 from connect.common.models import (
     Organization,
+    OrganizationRole,
     User,
     Project,
     OrganizationAuthorization,
     RequestPermissionProject,
     ProjectAuthorization,
-
 )
 from connect.usecases.organizations.retrieve import RetrieveOrganizationUseCase
 from connect.usecases.users.retrieve import RetrieveUserUseCase
 
-from connect.usecases.authorizations.dto import DeleteAuthorizationDTO, DeleteProjectAuthorizationDTO
+from connect.usecases.authorizations.dto import (
+    DeleteAuthorizationDTO,
+    DeleteProjectAuthorizationDTO,
+)
 
 from connect.usecases.authorizations.usecase import AuthorizationUseCase
 
@@ -21,6 +24,16 @@ class DeleteAuthorizationUseCase(AuthorizationUseCase):
     def delete_organization_authorization(self, user: User, org: Organization):
         try:
             authorization = org.authorizations.get(user=user)
+            if (
+                org.authorizations.exclude(uuid=authorization.uuid)
+                .filter(role=OrganizationRole.ADMIN.value)
+                .count()
+                == 0
+            ):
+                raise PermissionDenied(
+                    "There must be at least one admin in the organization"
+                )
+
             authorization.delete()
 
             if self.publish_message:
@@ -29,17 +42,24 @@ class DeleteAuthorizationUseCase(AuthorizationUseCase):
                     org_uuid=str(org.uuid),
                     user_email=user.email,
                     role=authorization.role,
-                    org_intelligence=org.inteligence_organization
+                    org_intelligence=org.inteligence_organization,
                 )
         except OrganizationAuthorization.DoesNotExist:
-            print(f"OrganizationAuthorization matching query does not exist: Org {org.uuid} User {user.email}")
+            print(
+                f"OrganizationAuthorization matching query does not exist: Org {org.uuid} User {user.email}"
+            )
 
-    def delete_project_authorization(self, project: Project, user: User, role: int = None):
+    def delete_project_authorization(
+        self, project: Project, user: User, role: int = None
+    ):
 
         authorization = project.project_authorizations.get(user=user)
         authorization.delete()
-        
-        if not ProjectAuthorization.objects.filter(user=user, organization_authorization=authorization.organization_authorization).exists():
+
+        if not ProjectAuthorization.objects.filter(
+            user=user,
+            organization_authorization=authorization.organization_authorization,
+        ).exists():
             self.delete_organization_authorization(user=user, org=project.organization)
 
         if not role:
@@ -55,38 +75,52 @@ class DeleteAuthorizationUseCase(AuthorizationUseCase):
 
     def delete_authorization(self, auth_dto: DeleteAuthorizationDTO):
         if auth_dto.request_user:
-            request_user : User = RetrieveUserUseCase().get_user_by_email(email=auth_dto.request_user)
+            request_user: User = RetrieveUserUseCase().get_user_by_email(
+                email=auth_dto.request_user
+            )
 
         if auth_dto.user_email:
-            user: User = RetrieveUserUseCase().get_user_by_email(email=auth_dto.user_email)
+            user: User = RetrieveUserUseCase().get_user_by_email(
+                email=auth_dto.user_email
+            )
         elif auth_dto.id:
             user: User = RetrieveUserUseCase().get_user_by_id(id=auth_dto.id)
 
-        org: Organization = RetrieveOrganizationUseCase().get_organization_by_uuid(org_uuid=auth_dto.org_uuid)
+        org: Organization = RetrieveOrganizationUseCase().get_organization_by_uuid(
+            org_uuid=auth_dto.org_uuid
+        )
 
         if not org.authorizations.filter(user=request_user).exists():
-            raise PermissionDenied("User does not have permission to perform this action")
+            raise PermissionDenied(
+                "User does not have permission to perform this action"
+            )
 
         org_auth = org.authorizations.get(user=user)
 
-        projects_uuids: QuerySet = user.project_authorizations_user.filter(organization_authorization__organization=org).values_list("project", flat=True)
+        projects_uuids: QuerySet = user.project_authorizations_user.filter(
+            organization_authorization__organization=org
+        ).values_list("project", flat=True)
 
         for project_uuid in projects_uuids:
             project = Project.objects.get(uuid=project_uuid)
             project_role = self.organization_permission_mapper.get(org_auth.role)
             self.delete_project_authorization(
-                project=project,
-                user=user,
-                role=project_role
+                project=project, user=user, role=project_role
             )
 
-        org_auth: OrganizationAuthorization = self.delete_organization_authorization(user=user, org=org)
+        org_auth: OrganizationAuthorization = self.delete_organization_authorization(
+            user=user, org=org
+        )
 
     def delete_single_project_permission(self, auth_dto: DeleteProjectAuthorizationDTO):
         project = Project.objects.get(uuid=auth_dto.project_uuid)
         try:
-            request_auth = RequestPermissionProject.objects.get(email=auth_dto.user_email, project=project)
+            request_auth = RequestPermissionProject.objects.get(
+                email=auth_dto.user_email, project=project
+            )
             request_auth.delete()
         except RequestPermissionProject.DoesNotExist:
-            user: User = RetrieveUserUseCase().get_user_by_email(email=auth_dto.user_email)
+            user: User = RetrieveUserUseCase().get_user_by_email(
+                email=auth_dto.user_email
+            )
             self.delete_project_authorization(project=project, user=user)
